3.1.0.0

+++ TxDOTAgent/TxDOTCadWriter.cpp
@@ -211,7 +211,7 @@ int TxDOT_WriteCADDataToFile (FILE *fp, IBroker* pBroker, const CGirderKey& gird
 
 	/* 1. SPAN NUMBER */
 	TCHAR	spanNumber[5+1];
-	_stprintf_s(spanNumber, sizeof(spanNumber)/sizeof(TCHAR), _T("%d"), LABEL_SPAN(segmentKey.groupIndex));
+	_stprintf_s(spanNumber, sizeof(spanNumber)/sizeof(TCHAR), _T("%d"), (int)LABEL_SPAN(segmentKey.groupIndex));
 
 	/* 1. GIRDER NUMBER */
 	TCHAR  beamNumber[5+1];
@@ -914,16 +914,16 @@ void TxDOTCadWriter::WriteFinalData(FILE *fp, bool isExtended, bool isIBeam, Int
    // lastly write any information
    if (m_OutCome==SectionMismatch || m_OutCome==SectionsNotSymmetrical)
    {
-	   _ftprintf(fp, _T("Warning: Irregular, Non-standard debonding increments used for beam %s in span %2d. Cannot write debonding information to TxDOT CAD format.\n"),LABEL_GIRDER(gdrIdx),LABEL_SPAN(spanIdx));
+	   _ftprintf(fp, _T("Warning: Irregular, Non-standard debonding increments used for beam %s in span %2d. Cannot write debonding information to TxDOT CAD format.\n"),LABEL_GIRDER(gdrIdx),(int)LABEL_SPAN(spanIdx));
    }
    else if (m_OutCome==TooManySections)
    {
-	   _ftprintf(fp, _T("Warning: The number of debonded sections exceeds ten for beam %s in span %2d. Cannot write debonding information to TxDOT CAD format.\n"),LABEL_GIRDER(gdrIdx),LABEL_SPAN(spanIdx));
+	   _ftprintf(fp, _T("Warning: The number of debonded sections exceeds ten for beam %s in span %2d. Cannot write debonding information to TxDOT CAD format.\n"),LABEL_GIRDER(gdrIdx),(int)LABEL_SPAN(spanIdx));
    }
    else if (m_OutCome==NonStandardSection)
    {
       Float64 spac = ::ConvertFromSysUnits(m_SectionSpacing , unitMeasure::Feet );
-	   _ftprintf(fp, _T("Warning: Non-standard debonding increment of %6.3f ft used  for beam %s in span %2d. \n"),spac,LABEL_GIRDER(gdrIdx),LABEL_SPAN(spanIdx));
+	   _ftprintf(fp, _T("Warning: Non-standard debonding increment of %6.3f ft used  for beam %s in span %2d. \n"),spac,LABEL_GIRDER(gdrIdx),(int)LABEL_SPAN(spanIdx));
    }
 }
 
@@ -1070,7 +1070,7 @@ int TxDOT_WriteDistributionFactorsToFile (FILE *fp, IBroker* pBroker, const CGir
                           &gR,  &gR1,  &gR2 );
 
 	TCHAR	spanNumber[5+1];
-	_stprintf_s(spanNumber, sizeof(spanNumber)/sizeof(TCHAR), _T("%d"), LABEL_SPAN(spanIdx));
+	_stprintf_s(spanNumber, sizeof(spanNumber)/sizeof(TCHAR), _T("%d"), (int)LABEL_SPAN(spanIdx));
 
 	TCHAR  beamNumber[5+1];
 	_stprintf_s(beamNumber, sizeof(beamNumber)/sizeof(TCHAR), _T("%s"), LABEL_GIRDER(gdrIdx));
